refactor(toolkit): move simple apis to shared helper package
Move the easier apis from cli package to the shared helper. Easier here means little or no dependencies on other apis.

Tests will be moved later to keep the diff size managable. Code is still tested as the tests are still running from `aws-cdk` package.
