+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/index.ts
@@ -1,10 +1,17 @@
 export * from './aws-auth';
 export * from './cloud-assembly';
 export * from './cloudformation';
+export * from './context';
 export * from './diff';
 export * from './io';
+export * from './notices';
 export * from './plugin';
 export * from './require-approval';
 export * from './resource-import';
-export * from './toolkit-error';
+export * from './rwlock';
 export * from './settings';
+export * from './stack-events';
+export * from './toolkit-error';
+export * from './work-graph';
+export * from './tree';
+export * from './tags';
