+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/stack-events/index.ts
@@ -1,3 +1,5 @@
 export * from './stack-activity-monitor';
 export * from './stack-event-poller';
 export * from './stack-status';
+
+export { StackProgressMonitor } from './stack-progress-monitor';
