+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/stack-events/stack-progress-monitor.ts
@@ -1,4 +1,3 @@
-
 import * as util from 'util';
 import type { StackProgress } from '@aws-cdk/tmp-toolkit-helpers';
 import type { StackEvent } from '@aws-sdk/client-cloudformation';
