+++ packages/@aws-cdk/toolkit-lib/lib/actions/diff/private/helpers.ts
@@ -5,9 +5,9 @@ import * as fs from 'fs-extra';
 import * as uuid from 'uuid';
 import type { ChangeSetDiffOptions, DiffOptions, LocalFileDiffOptions } from '..';
 import { DiffMethod } from '..';
-import type { Deployments, ResourcesToImport, SdkProvider } from '../../../api/aws-cdk';
+import type { Deployments, ResourcesToImport } from '../../../api/aws-cdk';
 import { createDiffChangeSet, ResourceMigrator } from '../../../api/aws-cdk';
-import type { IoHelper, StackCollection, TemplateInfo } from '../../../api/shared-private';
+import type { IoHelper, SdkProvider, StackCollection, TemplateInfo } from '../../../api/shared-private';
 import { IO, removeNonImportResources } from '../../../api/shared-private';
 import { PermissionChangeType, ToolkitError } from '../../../api/shared-public';
 import { deserializeStructure, formatErrorMessage } from '../../../private/util';
