+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/private/source-builder.ts
@@ -6,10 +6,9 @@ import { ContextAwareCloudAssembly } from './context-aware-source';
 import { execInChildProcess } from './exec';
 import { ExecutionEnvironment, assemblyFromDirectory } from './prepare-source';
 import type { ToolkitServices } from '../../../toolkit/private';
-import type { ILock } from '../../aws-cdk';
-import { Context, RWLock } from '../../aws-cdk';
 import { IO } from '../../io/private';
-import { Settings } from '../../shared-private';
+import type { ILock } from '../../shared-private';
+import { Context, RWLock, Settings } from '../../shared-private';
 import { ToolkitError, AssemblyError } from '../../shared-public';
 import type { AssemblyBuilder } from '../source-builder';
