+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/index.ts
@@ -1,4 +1,3 @@
 export * from './toolkit';
 export * from './non-interactive-io-host';
 export * from './types';
-export * from '../api/shared-public';
