+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/private/index.ts
@@ -1,8 +1,7 @@
 
-import type { SdkProvider } from '../../api/aws-cdk';
 import type { ICloudAssemblySource } from '../../api/cloud-assembly';
 import { CachedCloudAssemblySource, StackAssembly } from '../../api/cloud-assembly/private';
-import type { IoHelper } from '../../api/shared-private';
+import type { SdkProvider, IoHelper } from '../../api/shared-private';
 
 /**
  * Helper struct to pass internal services around.
