+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/types.ts
@@ -63,68 +63,3 @@ export interface Environment {
    */
   readonly region: string;
 }
-/**
- * Result interface for toolkit.deploy operation
- */
-export interface DeployResult {
-  /**
-   * List of stacks deployed by this operation
-   */
-  readonly stacks: DeployedStack[];
-}
-
-/**
- * Information about a deployed stack
- */
-export interface DeployedStack {
-  /**
-   * The name of the deployed stack
-   *
-   * A stack name is unique inside its environment, but not unique globally.
-   */
-  readonly stackName: string;
-
-  /**
-   * The environment where the stack was deployed
-   *
-   * This environment is always concrete, because even though the CDK app's
-   * stack may be region-agnostic, in order to be deployed it will have to have
-   * been specialized.
-   */
-  readonly environment: Environment;
-
-  /**
-   * Hierarchical identifier
-   *
-   * This uniquely identifies the stack inside the CDK app.
-   *
-   * In practice this will be the stack's construct path, but unfortunately the
-   * Cloud Assembly contract doesn't require or guarantee that.
-   */
-  readonly hierarchicalId: string;
-
-  /**
-   * The ARN of the deployed stack
-   */
-  readonly stackArn: string;
-
-  /**
-   * The outputs of the deployed CloudFormation stack
-   */
-  readonly outputs: { [key: string]: string };
-}
-
-/**
- * An environment, which is an (account, region) pair
- */
-export interface Environment {
-  /**
-   * The account number
-   */
-  readonly account: string;
-
-  /**
-   * The region number
-   */
-  readonly region: string;
-}
