+++ packages/@aws-cdk/toolkit-lib/tsconfig.dev.json
@@ -28,6 +28,7 @@
     "target": "es2022",
     "incremental": true,
     "skipLibCheck": true,
+    "rootDir": ".",
     "composite": true,
     "outDir": "lib"
   },
