+++ packages/aws-cdk/lib/api/work-graph.ts
@@ -1,0 +1,1 @@
+export * from '../../../@aws-cdk/tmp-toolkit-helpers/src/api/work-graph';
